refactor(review): separate selection preparation and normalization
