+++ packages/lexical-review/src/ReviewOperations.ts
@@ -4,67 +4,61 @@ import {
   $getRoot,
   $getSelection,
   $isElementNode,
-  $isParagraphNode,
   $isRangeSelection,
   $isTextNode,
-  type ElementNode,
   type LexicalEditor,
   type LexicalNode,
-  type ParagraphNode,
-  type PointType,
   type RangeSelection,
   type TextNode,
 } from "lexical";
 import { createProposalId, isValidProposalId } from "./ProposalIdentity";
 import {
-  $canReviewElementNodesBeMerged,
   $createReviewDeletionNode,
   $createReviewInsertionNode,
   $isReviewDeletionNode,
   $isReviewInsertionNode,
   ReviewDeletionNode,
-  ReviewElementNode,
   ReviewInsertionNode,
+  type ReviewElementNode,
 } from "./ReviewNodes";
-
-const SUPPORTED_TEXT_FORMAT_MASK = 0b1111;
-
-type PointSnapshot = Readonly<{
-  key: string;
-  offset: number;
-  type: "element" | "text";
-}>;
-
-export type ReviewIntentRefusalCode =
-  | "ambiguous-boundary"
-  | "deletion-target-unavailable"
-  | "invalid-proposal-id"
-  | "invalid-structural-target"
-  | "unsafe-proposal-intersection"
-  | "unsupported-formatting"
-  | "unsupported-input"
-  | "unsupported-proposal-edit"
-  | "unsupported-structure"
-  | "unsupported-target"
-  | "unsupported-transfer";
-
-export type ReviewIntentRefusal = Readonly<{
-  code: ReviewIntentRefusalCode;
-  message: string;
-  status: "refused";
-}>;
-
-export type ReviewIntentError = Readonly<{
-  cause: unknown;
-  code: string;
-  message: string;
-}>;
-
-export type ReviewIntentOutcome<T = void> =
-  | Readonly<{ status: "changed"; value: T }>
-  | Readonly<{ status: "unchanged"; value: T }>
-  | ReviewIntentRefusal
-  | Readonly<{ error: ReviewIntentError; status: "failed" }>;
+import {
+  changed,
+  unchanged,
+  refusal,
+  type Preparation,
+  type ReviewIntentOutcome,
+} from "./ReviewIntent";
+export type {
+  ReviewIntentRefusalCode,
+  ReviewIntentRefusal,
+  ReviewIntentError,
+  ReviewIntentOutcome,
+} from "./ReviewIntent";
+import {
+  isRootParagraph,
+  isReviewElementNode,
+  getChildIndex,
+  getTextChildren,
+  validateParagraphStructure,
+  isTextBoundary,
+  previousCharacterOffset,
+  nextCharacterOffset,
+  inspectSelection,
+  buildProposalMap,
+  buildProposalMapAroundPoint,
+  getProposalOffset,
+  buildProposalSpan,
+  buildAcceptedMap,
+  getAcceptedOffset,
+  buildAcceptedSpan,
+  getStartEntry,
+  getEndEntry,
+  type AcceptedPoint,
+  type ProposalPoint,
+  type ProposalMap,
+  type ProposalSpan,
+  type AcceptedSpan,
+} from "./ReviewSelectionPreparation";
 
 export type ReviewProposalIdFactory = () => string;
 
@@ -72,829 +66,7 @@ export type ReviewAuthoringOptions = Readonly<{
   proposalIdFactory?: ReviewProposalIdFactory;
 }>;
 
-type AcceptedPoint = Readonly<{
-  association: "accepted";
-  childIndex: number;
-  node: TextNode | null;
-  offset: number;
-  paragraph: ParagraphNode;
-}>;
-
-type ProposalPoint = Readonly<{
-  association: "proposal";
-  childIndex: number;
-  node: TextNode | null;
-  offset: number;
-  paragraph: ParagraphNode;
-  wrapper: ReviewElementNode;
-}>;
-
-type SelectionPoint = AcceptedPoint | ProposalPoint;
-
-type SelectionInspection = Readonly<{
-  anchor: SelectionPoint;
-  backward: boolean;
-  collapsed: boolean;
-  focus: SelectionPoint;
-  selection: RangeSelection;
-}>;
-
-type Preparation<T> =
-  Readonly<{ status: "ready"; value: T }> | ReviewIntentRefusal;
-
-type ProposalMapEntry = Readonly<{
-  end: number;
-  node: TextNode;
-  start: number;
-  wrapper: ReviewElementNode;
-}>;
-
-type ProposalMap = Readonly<{
-  entries: readonly ProposalMapEntry[];
-  paragraph: ParagraphNode;
-  total: number;
-  wrappers: readonly ReviewElementNode[];
-  proposalId: string;
-}>;
-
-type ProposalSpan = Readonly<{
-  end: number;
-  map: ProposalMap;
-  start: number;
-}>;
-
-type AcceptedMapEntry = Readonly<{
-  childIndex: number;
-  end: number;
-  node: TextNode;
-  start: number;
-}>;
-
-type AcceptedMap = Readonly<{
-  entries: readonly AcceptedMapEntry[];
-  paragraph: ParagraphNode;
-  total: number;
-}>;
-
-type AcceptedSpan = Readonly<{
-  end: number;
-  map: AcceptedMap;
-  start: number;
-}>;
-
-function refusal(
-  code: ReviewIntentRefusalCode,
-  message: string,
-): ReviewIntentRefusal {
-  return { code, message, status: "refused" };
-}
-
-function changed(): ReviewIntentOutcome {
-  return { status: "changed", value: undefined };
-}
-
-function unchanged(): ReviewIntentOutcome {
-  return { status: "unchanged", value: undefined };
-}
-
-function isRootParagraph(node: LexicalNode | null): node is ParagraphNode {
-  return $isParagraphNode(node) && node.getParent() === $getRoot();
-}
-
-function isReviewElementNode(
-  node: LexicalNode | null | undefined,
-): node is ReviewElementNode {
-  return $isReviewDeletionNode(node) || $isReviewInsertionNode(node);
-}
-
-type RootProposalContext = Readonly<{
-  paragraph: ParagraphNode;
-  wrapper: ReviewElementNode;
-}>;
-
-function getRootProposalContext(
-  node: LexicalNode | null | undefined,
-): RootProposalContext | null {
-  if (!isReviewElementNode(node)) {
-    return null;
-  }
-  const paragraph = node.getParent();
-  return isRootParagraph(paragraph) ? { paragraph, wrapper: node } : null;
-}
-
-function isSameProposalNode(
-  node: LexicalNode | null | undefined,
-  reference: ReviewElementNode,
-): node is ReviewElementNode {
-  return (
-    isReviewElementNode(node) && $canReviewElementNodesBeMerged(reference, node)
-  );
-}
-
-function snapshotPoint(point: PointType): PointSnapshot {
-  return { key: point.key, offset: point.offset, type: point.type };
-}
-
-function restorePointAfterReviewElementMerge(
-  point: PointType,
-  snapshot: PointSnapshot,
-  left: ReviewElementNode,
-  right: ReviewElementNode,
-  parent: ElementNode,
-  leftChildCount: number,
-  rightIndex: number,
-): void {
-  if (snapshot.type === "element" && snapshot.key === right.getKey()) {
-    point.set(left.getKey(), leftChildCount + snapshot.offset, "element");
-    return;
-  }
-  if (snapshot.type === "element" && snapshot.key === parent.getKey()) {
-    if (snapshot.offset === rightIndex) {
-      point.set(left.getKey(), leftChildCount, "element");
-    } else if (snapshot.offset > rightIndex) {
-      point.set(parent.getKey(), snapshot.offset - 1, "element");
-    } else {
-      point.set(parent.getKey(), snapshot.offset, "element");
-    }
-    return;
-  }
-  point.set(snapshot.key, snapshot.offset, snapshot.type);
-}
-
-function mergeReviewElementNodes(
-  left: ReviewElementNode,
-  right: ReviewElementNode,
-): void {
-  const parent = left.getParent();
-  const rightIndex = right.getIndexWithinParent();
-  if (!$isElementNode(parent) || rightIndex < 0) {
-    return;
-  }
-  const leftChildCount = left.getChildrenSize();
-  const selection = $getSelection();
-  const anchor = $isRangeSelection(selection)
-    ? snapshotPoint(selection.anchor)
-    : null;
-  const focus = $isRangeSelection(selection)
-    ? snapshotPoint(selection.focus)
-    : null;
-
-  left.append(...right.getChildren());
-  right.remove();
-
-  if ($isRangeSelection(selection) && anchor !== null && focus !== null) {
-    restorePointAfterReviewElementMerge(
-      selection.anchor,
-      anchor,
-      left,
-      right,
-      parent,
-      leftChildCount,
-      rightIndex,
-    );
-    restorePointAfterReviewElementMerge(
-      selection.focus,
-      focus,
-      left,
-      right,
-      parent,
-      leftChildCount,
-      rightIndex,
-    );
-    selection.dirty = true;
-  }
-}
-
-export function normalizeReviewElementNode(node: ReviewElementNode): void {
-  if (!node.isAttached()) {
-    return;
-  }
-  const previous = node.getPreviousSibling();
-  if (
-    isReviewElementNode(previous) &&
-    $canReviewElementNodesBeMerged(previous, node)
-  ) {
-    mergeReviewElementNodes(previous, node);
-    return;
-  }
-  const next = node.getNextSibling();
-  if (isReviewElementNode(next) && $canReviewElementNodesBeMerged(node, next)) {
-    mergeReviewElementNodes(node, next);
-  }
-}
-
-function getChildIndex(parent: ElementNode, node: LexicalNode): number | null {
-  const index = parent
-    .getChildren()
-    .findIndex((child) => child.getKey() === node.getKey());
-  return index === -1 ? null : index;
-}
-
-function getTextChildren(wrapper: ReviewElementNode): TextNode[] | null {
-  const children = wrapper.getChildren();
-  if (
-    children.length === 0 ||
-    children.some(
-      (child) => !$isTextNode(child) || child.getTextContentSize() === 0,
-    )
-  ) {
-    return null;
-  }
-  return children.filter($isTextNode);
-}
-
-function validateParagraphStructure(
-  paragraph: ParagraphNode,
-): ReviewIntentRefusal | null {
-  for (const child of paragraph.getChildren()) {
-    if ($isTextNode(child)) {
-      if (hasUnsupportedTextFormatting(child)) {
-        return refusal(
-          "unsupported-formatting",
-          "Review editing supports bold, italic, strikethrough, and underline text without inline styles or token modes.",
-        );
-      }
-      continue;
-    }
-    if (isReviewElementNode(child)) {
-      const textChildren = getTextChildren(child);
-      if (textChildren === null) {
-        return refusal(
-          "invalid-structural-target",
-          "Review editing supports only direct paragraph text and text-only proposal wrappers.",
-        );
-      }
-      if (textChildren.some(hasUnsupportedTextFormatting)) {
-        return refusal(
-          "unsupported-formatting",
-          "Review editing supports bold, italic, strikethrough, and underline text without inline styles or token modes.",
-        );
-      }
-      continue;
-    }
-    return refusal(
-      "invalid-structural-target",
-      "Review editing supports only direct paragraph text and text-only proposal wrappers.",
-    );
-  }
-  return null;
-}
-
-function hasUnsupportedTextFormatting(node: TextNode): boolean {
-  return (
-    (node.getFormat() & ~SUPPORTED_TEXT_FORMAT_MASK) !== 0 ||
-    node.getDetail() !== 0 ||
-    node.getMode() !== "normal" ||
-    node.getStyle() !== ""
-  );
-}
-
-function validateSelectionFormatting(
-  selection: RangeSelection,
-): ReviewIntentRefusal | null {
-  return (selection.format & ~SUPPORTED_TEXT_FORMAT_MASK) !== 0 ||
-    selection.style !== ""
-    ? refusal(
-        "unsupported-formatting",
-        "Review editing supports bold, italic, strikethrough, and underline text without inline styles.",
-      )
-    : null;
-}
-
-function isTextBoundary(text: string, offset: number): boolean {
-  if (offset <= 0 || offset >= text.length) {
-    return true;
-  }
-  const before = text.charCodeAt(offset - 1);
-  const after = text.charCodeAt(offset);
-  return !(
-    before >= 0xd800 &&
-    before <= 0xdbff &&
-    after >= 0xdc00 &&
-    after <= 0xdfff
-  );
-}
-
-function previousCharacterOffset(text: string, offset: number): number {
-  const previous = offset - 1;
-  if (
-    previous > 0 &&
-    text.charCodeAt(previous) >= 0xdc00 &&
-    text.charCodeAt(previous) <= 0xdfff &&
-    text.charCodeAt(previous - 1) >= 0xd800 &&
-    text.charCodeAt(previous - 1) <= 0xdbff
-  ) {
-    return previous - 1;
-  }
-  return previous;
-}
-
-function nextCharacterOffset(text: string, offset: number): number {
-  const next = offset + 1;
-  if (
-    next < text.length &&
-    text.charCodeAt(offset) >= 0xd800 &&
-    text.charCodeAt(offset) <= 0xdbff &&
-    text.charCodeAt(next) >= 0xdc00 &&
-    text.charCodeAt(next) <= 0xdfff
-  ) {
-    return next + 1;
-  }
-  return next;
-}
-
-function classifyPoint(point: PointType): Preparation<SelectionPoint> {
-  const node = point.getNode();
-  if (point.type === "text") {
-    if (!$isTextNode(node)) {
-      return refusal(
-        "invalid-structural-target",
-        "A text selection point must identify a Lexical text node.",
-      );
-    }
-    const text = node.getTextContent();
-    if (
-      !Number.isInteger(point.offset) ||
-      point.offset < 0 ||
-      point.offset > text.length ||
-      !isTextBoundary(text, point.offset)
-    ) {
-      return refusal(
-        "invalid-structural-target",
-        "The text selection point is outside a supported Unicode text boundary.",
-      );
-    }
-    const parentNode: LexicalNode | null = node.getParent();
-    if (isRootParagraph(parentNode)) {
-      const structure = validateParagraphStructure(parentNode);
-      if (structure !== null) {
-        return structure;
-      }
-      const childIndex = getChildIndex(parentNode, node);
-      if (childIndex === null) {
-        return refusal(
-          "invalid-structural-target",
-          "The selected text node is not attached to its paragraph.",
-        );
-      }
-      return {
-        status: "ready",
-        value: {
-          association: "accepted",
-          childIndex,
-          node,
-          offset: point.offset,
-          paragraph: parentNode,
-        },
-      };
-    }
-    const proposal = getRootProposalContext(parentNode);
-    if (proposal !== null) {
-      const { paragraph, wrapper: parent } = proposal;
-      const structure = validateParagraphStructure(paragraph);
-      if (structure !== null) {
-        return structure;
-      }
-      const childIndex = getChildIndex(paragraph, parent);
-      if (childIndex === null) {
-        return refusal(
-          "invalid-structural-target",
-          "The selected proposal wrapper is not attached to its paragraph.",
-        );
-      }
-      const textChildren = getTextChildren(parent);
-      if (textChildren === null) {
-        return refusal(
-          "invalid-structural-target",
-          "The selected proposal wrapper has unsupported live children.",
-        );
-      }
-      return {
-        status: "ready",
-        value: {
-          association: "proposal",
-          childIndex,
-          node,
-          offset: point.offset,
-          paragraph,
-          wrapper: parent,
-        },
-      };
-    }
-    return refusal(
-      "invalid-structural-target",
-      "Review editing supports only direct paragraph text and proposal text.",
-    );
-  }
-
-  if (!$isElementNode(node)) {
-    return refusal(
-      "invalid-structural-target",
-      "An element selection point must identify a Lexical element node.",
-    );
-  }
-  if (!Number.isInteger(point.offset) || point.offset < 0) {
-    return refusal(
-      "invalid-structural-target",
-      "The element selection point has an invalid child offset.",
-    );
-  }
-  const proposal = getRootProposalContext(node);
-  if (proposal !== null) {
-    const { paragraph, wrapper } = proposal;
-    const textChildren = getTextChildren(wrapper);
-    if (textChildren === null || point.offset > textChildren.length) {
-      return refusal(
-        "invalid-structural-target",
-        "The proposal element point does not identify a supported child boundary.",
-      );
-    }
-    const structure = validateParagraphStructure(paragraph);
-    if (structure !== null) {
-      return structure;
-    }
-    const childIndex = getChildIndex(paragraph, node);
-    if (childIndex === null) {
-      return refusal(
-        "invalid-structural-target",
-        "The selected proposal wrapper is not attached to its paragraph.",
-      );
-    }
-    return {
-      status: "ready",
-      value: {
-        association: "proposal",
-        childIndex,
-        node: null,
-        offset: textChildren
-          .slice(0, point.offset)
-          .reduce((total, child) => total + child.getTextContentSize(), 0),
-        paragraph,
-        wrapper,
-      },
-    };
-  }
-  if (isRootParagraph(node)) {
-    const structure = validateParagraphStructure(node);
-    if (structure !== null) {
-      return structure;
-    }
-    const children = node.getChildren();
-    if (point.offset > children.length) {
-      return refusal(
-        "invalid-structural-target",
-        "The paragraph element point is outside its child range.",
-      );
-    }
-    const left = children[point.offset - 1];
-    const right = children[point.offset];
-    if (isReviewElementNode(left) || isReviewElementNode(right)) {
-      return refusal(
-        "ambiguous-boundary",
-        "A paragraph boundary next to proposal content does not identify one editing side.",
-      );
-    }
-    return {
-      status: "ready",
-      value: {
-        association: "accepted",
-        childIndex: point.offset,
-        node: null,
-        offset: children
-          .slice(0, point.offset)
-          .reduce((total, child) => total + child.getTextContentSize(), 0),
-        paragraph: node,
-      },
-    };
-  }
-  return refusal(
-    "invalid-structural-target",
-    "Review editing supports only paragraph and proposal element points.",
-  );
-}
-
-function inspectSelection(): Preparation<SelectionInspection> {
-  const selection = $getSelection();
-  if (!$isRangeSelection(selection)) {
-    return refusal(
-      "unsupported-target",
-      "Review editing requires one Lexical range selection.",
-    );
-  }
-  const formatting = validateSelectionFormatting(selection);
-  if (formatting !== null) {
-    return formatting;
-  }
-  const anchor = classifyPoint(selection.anchor);
-  if (anchor.status !== "ready") {
-    return anchor;
-  }
-  const focus = classifyPoint(selection.focus);
-  if (focus.status !== "ready") {
-    return focus;
-  }
-  return {
-    status: "ready",
-    value: {
-      anchor: anchor.value,
-      backward: selection.isBackward(),
-      collapsed: selection.isCollapsed(),
-      focus: focus.value,
-      selection,
-    },
-  };
-}
-
-function sameProposal(left: ProposalPoint, right: ProposalPoint): boolean {
-  return (
-    left.paragraph === right.paragraph &&
-    isSameProposalNode(left.wrapper, right.wrapper)
-  );
-}
-
-function buildProposalMap(
-  paragraph: ParagraphNode,
-  startWrapper: ReviewElementNode,
-  endWrapper: ReviewElementNode,
-): Preparation<ProposalMap> {
-  const startIndex = getChildIndex(paragraph, startWrapper);
-  const endIndex = getChildIndex(paragraph, endWrapper);
-  if (startIndex === null || endIndex === null || startIndex > endIndex) {
-    return refusal(
-      "invalid-structural-target",
-      "The proposal selection wrappers are not ordered in one paragraph.",
-    );
-  }
-  const proposalId = startWrapper.getProposalId();
-  const wrappers: ReviewElementNode[] = [];
-  const entries: ProposalMapEntry[] = [];
-  let offset = 0;
-  const children = paragraph.getChildren();
-  for (let index = startIndex; index <= endIndex; index += 1) {
-    const child = children[index];
-    if (!isSameProposalNode(child, startWrapper)) {
-      return refusal(
-        "unsafe-proposal-intersection",
-        "The selection intersects accepted content or another proposal identity.",
-      );
-    }
-    const textChildren = getTextChildren(child);
-    if (textChildren === null) {
-      return refusal(
-        "invalid-structural-target",
-        "A pending proposal contains unsupported live children.",
-      );
-    }
-    wrappers.push(child);
-    for (const node of textChildren) {
-      const end = offset + node.getTextContentSize();
-      entries.push({ end, node, start: offset, wrapper: child });
-      offset = end;
-    }
-  }
-  if (entries.length === 0) {
-    return refusal(
-      "invalid-structural-target",
-      "A pending proposal must contain live text before it can be edited.",
-    );
-  }
-  return {
-    status: "ready",
-    value: {
-      entries,
-      paragraph,
-      proposalId,
-      total: offset,
-      wrappers,
-    },
-  };
-}
-
-function buildProposalMapAroundPoint(
-  point: ProposalPoint,
-): Preparation<ProposalMap> {
-  const children = point.paragraph.getChildren();
-  let startIndex = point.childIndex;
-  let endIndex = point.childIndex;
-  const isSameWrapper = (
-    child: LexicalNode | undefined,
-  ): child is ReviewElementNode => isSameProposalNode(child, point.wrapper);
-
-  while (startIndex > 0 && isSameWrapper(children[startIndex - 1])) {
-    startIndex -= 1;
-  }
-  while (
-    endIndex + 1 < children.length &&
-    isSameWrapper(children[endIndex + 1])
-  ) {
-    endIndex += 1;
-  }
-  const startWrapper = children[startIndex];
-  const endWrapper = children[endIndex];
-  if (!isSameWrapper(startWrapper) || !isSameWrapper(endWrapper)) {
-    return refusal(
-      "invalid-structural-target",
-      "The proposal caret is not attached to a supported proposal run.",
-    );
-  }
-  return buildProposalMap(point.paragraph, startWrapper, endWrapper);
-}
-
-function getProposalOffset(
-  point: ProposalPoint,
-  map: ProposalMap,
-): number | null {
-  let offset = 0;
-  for (const wrapper of map.wrappers) {
-    const children = getTextChildren(wrapper);
-    if (children === null) {
-      return null;
-    }
-    if (wrapper.getKey() === point.wrapper.getKey()) {
-      if (point.node === null) {
-        return offset + point.offset;
-      }
-      for (const child of children) {
-        if (child.getKey() === point.node.getKey()) {
-          return offset + point.offset;
-        }
-        offset += child.getTextContentSize();
-      }
-      return null;
-    }
-    offset += children.reduce(
-      (total, child) => total + child.getTextContentSize(),
-      0,
-    );
-  }
-  return null;
-}
-
-function buildProposalSpan(
-  inspection: SelectionInspection,
-): Preparation<ProposalSpan> {
-  if (
-    inspection.anchor.association !== "proposal" ||
-    inspection.focus.association !== "proposal"
-  ) {
-    return refusal(
-      "unsafe-proposal-intersection",
-      "The selection does not stay on one proposal side.",
-    );
-  }
-  if (!sameProposal(inspection.anchor, inspection.focus)) {
-    return refusal(
-      "unsafe-proposal-intersection",
-      "A selection may edit only one proposal identity and kind at a time.",
-    );
-  }
-  const startPoint = inspection.backward ? inspection.focus : inspection.anchor;
-  const endPoint = inspection.backward ? inspection.anchor : inspection.focus;
-  const map = buildProposalMap(
-    startPoint.paragraph,
-    startPoint.wrapper,
-    endPoint.wrapper,
-  );
-  if (map.status !== "ready") {
-    return map;
-  }
-  const start = getProposalOffset(startPoint, map.value);
-  const end = getProposalOffset(endPoint, map.value);
-  if (start === null || end === null || end < start) {
-    return refusal(
-      "invalid-structural-target",
-      "The proposal selection points cannot be ordered in the live tree.",
-    );
-  }
-  return { status: "ready", value: { end, map: map.value, start } };
-}
-
-function buildAcceptedMap(paragraph: ParagraphNode): Preparation<AcceptedMap> {
-  const entries: AcceptedMapEntry[] = [];
-  let offset = 0;
-  for (const [childIndex, child] of paragraph.getChildren().entries()) {
-    if (!$isTextNode(child)) {
-      continue;
-    }
-    const end = offset + child.getTextContentSize();
-    if (child.getTextContentSize() > 0) {
-      entries.push({ childIndex, end, node: child, start: offset });
-    }
-    offset = end;
-  }
-  return {
-    status: "ready",
-    value: { entries, paragraph, total: offset },
-  };
-}
-
-function getAcceptedOffset(
-  point: AcceptedPoint,
-  map: AcceptedMap,
-): number | null {
-  let offset = 0;
-  const children = map.paragraph.getChildren();
-  for (let childIndex = 0; childIndex < children.length; childIndex += 1) {
-    if (childIndex > point.childIndex) {
-      break;
-    }
-    const child = children[childIndex];
-    if (child === undefined) {
-      return null;
-    }
-    if (childIndex === point.childIndex) {
-      if (point.node === null) {
-        return offset;
-      }
-      if (!$isTextNode(child) || child.getKey() !== point.node.getKey()) {
-        return null;
-      }
-      return offset + point.offset;
-    }
-    if ($isTextNode(child)) {
-      offset += child.getTextContentSize();
-    }
-  }
-  return point.node === null &&
-    point.childIndex === map.paragraph.getChildrenSize()
-    ? offset
-    : null;
-}
-
-function buildAcceptedSpan(
-  inspection: SelectionInspection,
-): Preparation<AcceptedSpan> {
-  if (
-    inspection.anchor.association !== "accepted" ||
-    inspection.focus.association !== "accepted"
-  ) {
-    return refusal(
-      "unsafe-proposal-intersection",
-      "The selection intersects proposal-side content.",
-    );
-  }
-  if (inspection.anchor.paragraph !== inspection.focus.paragraph) {
-    return refusal(
-      "unsupported-target",
-      "Accepted editing supports one same-paragraph range.",
-    );
-  }
-  const startPoint = inspection.backward ? inspection.focus : inspection.anchor;
-  const endPoint = inspection.backward ? inspection.anchor : inspection.focus;
-  const startIndex = startPoint.childIndex;
-  const endIndex = endPoint.node
-    ? endPoint.childIndex
-    : endPoint.childIndex - 1;
-  if (startIndex > endIndex + 1) {
-    return refusal(
-      "invalid-structural-target",
-      "The accepted selection points are not ordered in the paragraph.",
-    );
-  }
-  const children = startPoint.paragraph.getChildren();
-  for (let index = startIndex; index <= endIndex; index += 1) {
-    const child = children[index];
-    if (child !== undefined && !$isTextNode(child)) {
-      return refusal(
-        "unsafe-proposal-intersection",
-        "The accepted range crosses pending proposal content.",
-      );
-    }
-  }
-  const map = buildAcceptedMap(startPoint.paragraph);
-  if (map.status !== "ready") {
-    return map;
-  }
-  const start = getAcceptedOffset(startPoint, map.value);
-  const end = getAcceptedOffset(endPoint, map.value);
-  if (start === null || end === null || end < start) {
-    return refusal(
-      "invalid-structural-target",
-      "The accepted selection points cannot be resolved in the live tree.",
-    );
-  }
-  return { status: "ready", value: { end, map: map.value, start } };
-}
-
-function getStartEntry(
-  entries: readonly AcceptedMapEntry[] | readonly ProposalMapEntry[],
-  offset: number,
-): AcceptedMapEntry | ProposalMapEntry | null {
-  return (
-    entries.find((entry) => entry.start <= offset && offset < entry.end) ?? null
-  );
-}
-
-function getEndEntry(
-  entries: readonly AcceptedMapEntry[] | readonly ProposalMapEntry[],
-  offset: number,
-): AcceptedMapEntry | ProposalMapEntry | null {
-  return (
-    entries.find((entry) => entry.start < offset && offset <= entry.end) ?? null
-  );
-}
-
-function getAcceptedSelectedNodes(span: AcceptedSpan): TextNode[] | null {
+function isolateAcceptedTextRange(span: AcceptedSpan): TextNode[] | null {
   if (span.start >= span.end) {
     return null;
   }
@@ -966,24 +138,35 @@ function placeProposalCaret(
   );
 }
 
-function removeProposalRange(
+function spliceProposalRange(
   span: ProposalSpan,
-  start: number,
-  end: number,
+  replacement: Readonly<{ node: TextNode; text: string }> | null = null,
 ): void {
   for (let index = span.map.entries.length - 1; index >= 0; index -= 1) {
     const entry = span.map.entries[index];
     if (entry === undefined) {
       continue;
     }
-    const localStart = Math.max(start, entry.start) - entry.start;
-    const localEnd = Math.min(end, entry.end) - entry.start;
+    const localStart = Math.max(span.start, entry.start) - entry.start;
+    const localEnd = Math.min(span.end, entry.end) - entry.start;
     if (localStart >= localEnd) {
       continue;
     }
-    entry.node.spliceText(localStart, localEnd - localStart, "", false);
-    if (entry.node.getTextContentSize() === 0) {
-      entry.node.remove();
+    if (
+      replacement !== null &&
+      entry.node.getKey() === replacement.node.getKey()
+    ) {
+      entry.node.spliceText(
+        localStart,
+        localEnd - localStart,
+        replacement.text,
+        true,
+      );
+    } else {
+      entry.node.spliceText(localStart, localEnd - localStart, "", false);
+      if (entry.node.getTextContentSize() === 0) {
+        entry.node.remove();
+      }
     }
   }
   for (const wrapper of span.map.wrappers) {
@@ -1012,30 +195,7 @@ function replaceProposalRange(
     span.map.paragraph,
     span.map.wrappers[0]!,
   );
-  for (let index = span.map.entries.length - 1; index >= 0; index -= 1) {
-    const entry = span.map.entries[index];
-    if (entry === undefined) {
-      continue;
-    }
-    const localStart = Math.max(span.start, entry.start) - entry.start;
-    const localEnd = Math.min(span.end, entry.end) - entry.start;
-    if (localStart >= localEnd) {
-      continue;
-    }
-    if (entry.node.getKey() === startEntry.node.getKey()) {
-      entry.node.spliceText(localStart, localEnd - localStart, text, true);
-    } else {
-      entry.node.spliceText(localStart, localEnd - localStart, "", false);
-      if (entry.node.getTextContentSize() === 0) {
-        entry.node.remove();
-      }
-    }
-  }
-  for (const wrapper of span.map.wrappers) {
-    if (wrapper.getChildrenSize() === 0) {
-      wrapper.remove();
-    }
-  }
+  spliceProposalRange(span, { node: startEntry.node, text });
   placeProposalCaret(span.map, span.start + text.length, fallbackIndex ?? 0);
   return changed();
 }
@@ -1117,17 +277,13 @@ function missingProposalNode(
       );
 }
 
-function insertAcceptedProposal(
+function insertInsertionProposalAtAcceptedPoint(
   point: AcceptedPoint,
   selection: RangeSelection,
-  kind: "deletion" | "insertion",
   proposalId: string,
   text: string,
 ): void {
-  const wrapper =
-    kind === "insertion"
-      ? $createReviewInsertionNode(proposalId)
-      : $createReviewDeletionNode(proposalId);
+  const wrapper = $createReviewInsertionNode(proposalId);
   const textNode = $createTextNode(text);
   textNode.setFormat(point.node?.getFormat() ?? selection.format);
   wrapper.append(textNode);
@@ -1154,12 +310,9 @@ function acceptedDeletionTarget(
   point: AcceptedPoint,
   backward: boolean,
   granularity: "character" | "word",
-): Readonly<{ end: number; map: AcceptedMap; start: number }> | null {
+): AcceptedSpan | null {
   const map = buildAcceptedMap(point.paragraph);
-  if (map.status !== "ready") {
-    return null;
-  }
-  const offset = getAcceptedOffset(point, map.value);
+  const offset = getAcceptedOffset(point, map);
   if (offset === null) {
     return null;
   }
@@ -1172,7 +325,7 @@ function acceptedDeletionTarget(
     let right = index;
     while ($isTextNode(children[left - 1])) left -= 1;
     while ($isTextNode(children[right + 1])) right += 1;
-    const entries = map.value.entries.filter(
+    const entries = map.entries.filter(
       (entry) => entry.childIndex >= left && entry.childIndex <= right,
     );
     const base = entries[0]?.start;
@@ -1183,7 +336,7 @@ function acceptedDeletionTarget(
     return {
       start: base + Math.min(local, boundary),
       end: base + Math.max(local, boundary),
-      map: map.value,
+      map: map,
     };
   }
   if (point.node !== null) {
@@ -1191,7 +344,7 @@ function acceptedDeletionTarget(
     if (backward && point.offset > 0) {
       return {
         end: offset,
-        map: map.value,
+        map: map,
         start:
           offset - (point.offset - previousCharacterOffset(text, point.offset)),
       };
@@ -1199,7 +352,7 @@ function acceptedDeletionTarget(
     if (!backward && point.offset < text.length) {
       return {
         end: offset + (nextCharacterOffset(text, point.offset) - point.offset),
-        map: map.value,
+        map: map,
         start: offset,
       };
     }
@@ -1213,7 +366,7 @@ function acceptedDeletionTarget(
   if (!$isTextNode(adjacent) || adjacent.getTextContentSize() === 0) {
     return null;
   }
-  const adjacentEntry = map.value.entries.find(
+  const adjacentEntry = map.entries.find(
     (entry) => entry.node.getKey() === adjacent.getKey(),
   );
   if (adjacentEntry === undefined) {
@@ -1226,14 +379,14 @@ function acceptedDeletionTarget(
     );
     return {
       end: adjacentEntry.end,
-      map: map.value,
+      map: map,
       start: adjacentEntry.start + start,
     };
   }
   const end = nextCharacterOffset(adjacent.getTextContent(), 0);
   return {
     end: adjacentEntry.start + end,
-    map: map.value,
+    map: map,
     start: adjacentEntry.start,
   };
 }
@@ -1282,7 +435,7 @@ function deleteProposalAtCaret(
   const end = Math.max(offset, boundary);
   const span: ProposalSpan = { end, map: map.value, start };
   const fallbackIndex = point.childIndex;
-  removeProposalRange(span, start, end);
+  spliceProposalRange(span);
   placeProposalCaret(map.value, start, fallbackIndex);
   return changed();
 }
@@ -1297,7 +450,7 @@ function deleteProposalSelection(span: ProposalSpan): ReviewIntentOutcome {
     span.map.paragraph,
     span.map.wrappers[0]!,
   );
-  removeProposalRange(span, span.start, span.end);
+  spliceProposalRange(span);
   placeProposalCaret(span.map, span.start, fallbackIndex ?? 0);
   return changed();
 }
@@ -1404,10 +557,9 @@ function performInsertion(
   if (proposalId.status !== "ready") {
     return proposalId;
   }
-  insertAcceptedProposal(
+  insertInsertionProposalAtAcceptedPoint(
     point,
     inspection.value.selection,
-    "insertion",
     proposalId.value,
     text,
   );
@@ -1677,7 +829,7 @@ function deleteAcceptedSpan(
       ? prepareProposalId(options)
       : { status: "ready" as const, value: continuation.getProposalId() };
   if (identity.status !== "ready") return identity;
-  const selected = getAcceptedSelectedNodes(span);
+  const selected = isolateAcceptedTextRange(span);
   if (selected === null || selected.length === 0)
     throw new Error("Validated deletion target could not be isolated.");
   const wrapper = continuation ?? $createReviewDeletionNode(identity.value);
